hover highlight refactor
